+++ src/utils.js
@@ -70,6 +70,8 @@ export function canCreateNewRequest(item, stripes) {
     includes([
       itemStatusesMap.IN_PROCESS,
       itemStatusesMap.AVAILABLE,
+      itemStatusesMap.CHECKED_OUT,
+      itemStatusesMap.ON_ORDER,
       itemStatusesMap.IN_TRANSIT,
       itemStatusesMap.AWAITING_PICKUP,
       itemStatusesMap.MISSING,